portfolio_project/urls.py
from django.contrib import admin
from django.urls import path, include
from django.conf import settings
from django.conf.urls.static import static
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView, TokenVerifyView

urlpatterns = [
    path('admin/', admin.site.urls),

    # App URLs
    path('', include('portfolio.urls')),
    path('projects/', include('projects.urls')),
    path('contact/', include('contact.urls')),
    path('accounts/', include('accounts.urls')),

    # AllAuth (OAuth)
    path('accounts/', include('allauth.urls')),

    # JWT API endpoints
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/token/verify/', TokenVerifyView.as_view(), name='token_verify'),

    # REST API
    path('api/', include('portfolio.api_urls')),
    path('api/projects/', include('projects.api_urls')),
    path('api/contact/', include('contact.api_urls')),
]

if settings.DEBUG:
    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
    urlpatterns += static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
