requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-allauth>=0.57
django-cors-headers>=4.3
Pillow>=10.0
whitenoise>=6.6
python-dotenv>=1.0
gunicorn>=21.2
